Make prior params buffers, elegant _apply override
Priors are now also Modules, attached as children to the module to which
the parameter the prior is over belongs. I think this is pretty clean,
and the pytorch representation *i.e. just call `model`) makes things very
clear as to which parameters have priors.
I also found a reasonably elegant way to overwrite _apply in the priors
in a mininmally verbose fashion. Basically, all parameters and buffers
are updated using the pytorch internals, but the distributions that live
in the prior objects are re-initialized upon call of _apply. This makes
it such that calling `model.cuda()` on the top level model will just
work.

There is still more work to be done - lots of test failures in the
examples, some collision of `Module.initialize()` with
`SpectralMixtureKernel.initialize()`, and the need to explicitly pass in
the size/dimension to the priors we already have in the examples.